Rename test package to not have RestVerticle interfere with tests Looks like RestVerticle instantiates all org.folio.rest.impl.* classes, which causes trouble during tests
