+++ mod-erm-usage-server/src/test/java/org/folio/rest/impl2/AggregatorSettingsExportIT.java
@@ -1,4 +1,4 @@
-package org.folio.rest.impl;
+package org.folio.rest.impl2;
 
 import static io.restassured.RestAssured.given;
 import static org.assertj.core.api.Assertions.assertThat;
