+++ mod-erm-usage-server/src/test/java/org/folio/rest/impl2/AggregatorSettingsTest.java
@@ -1,4 +1,4 @@
-package org.folio.rest.impl;
+package org.folio.rest.impl2;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -8,6 +8,7 @@ import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Collections;
 import java.util.Date;
+import org.folio.rest.impl.AggregatorSettingsAPI;
 import org.folio.rest.jaxrs.model.Metadata;
 import org.folio.rest.jaxrs.model.UsageDataProvider;
 import org.junit.Test;
