+++ mod-erm-usage-server/src/test/java/org/folio/rest/impl2/CounterReportUploadIT.java
@@ -1,4 +1,4 @@
-package org.folio.rest.impl;
+package org.folio.rest.impl2;
 
 import static io.restassured.RestAssured.get;
 import static io.restassured.RestAssured.given;
@@ -34,6 +34,7 @@ import javax.xml.bind.JAXB;
 import org.folio.okapi.common.XOkapiHeaders;
 import org.folio.postgres.testing.PostgresTesterContainer;
 import org.folio.rest.RestVerticle;
+import org.folio.rest.impl.TenantAPI;
 import org.folio.rest.jaxrs.model.Contents;
 import org.folio.rest.jaxrs.model.CounterReport;
 import org.folio.rest.jaxrs.model.CounterReportDocument;
